(less) Rework navbar padding and margins for right aligned last-child elements
This slight reworking of the navbar spacing was brought about by a
problem noted in #11530: use of `.navbar-form.navbar-right` leaves an
extra 15px of space on the right.

Here’s what’s new:

* Removed the `padding` on `.navbar-collapse` within static top, fixed
top, and fixed bottom navbars. Meaning, it’s still there for the
default navbar. This was the root cause of the extra 15px padding for
navbar forms.
* As part of this, I also replaced the existing negative margin
overrides for last-child elements with more contextual ones within
their sub-components.

Overall not too different, but should be more effective now and easier
to work with now.

For example of new changes, see http://jsbin.com/odeSiSo/1/.